chore: move packages to internal folder
