+++ internal/inventory/account_test.go
