+++ internal/inventory/cluster.go
