+++ internal/inventory/instance.go
