+++ internal/inventory/types.go
