+++ internal/stocker/stocker.go
@@ -1,6 +1,6 @@
 package stocker
 
-import "github.com/RHEcosystemAppEng/cluster-iq/pkg/inventory"
+import "github.com/RHEcosystemAppEng/cluster-iq/internal/inventory"
 
 // Stocker interface
 type Stocker interface {
